🚚 Move Changelog Page to Pro
