+++ lib/changelog.js
@@ -14,9 +14,9 @@ const getLogs = async () => {
 
       return {
         title: document.data.title,
-        project: document.data.project,
         author: document.data.author,
         date: document.data.date,
+        color: document.data.color,
         markdown: document.content,
       }
     })
