--- pages/changelog.module.css
@@ -1,222 +1,0 @@
-.changelog {
-    box-sizing: border-box;
-    line-height: 1.5;
-
-    width: 100%;
-    min-height: 100vh;
-    position: relative;
-    overflow: hidden;
-    display: grid;
-    grid-auto-flow: row;
-    align-content: flex-start;
-    align-items: flex-start;
-    gap: 2rem;
-    font-family: 'Fira Code', monospace;
-    background: hsl(245, 14%, 15%);
-}
-
-.container {
-    width: 64rem;
-    max-width: 100%;
-    margin: 0 auto;
-}
-
-.header {
-    padding: 10rem 10% 6rem 10%;
-    background-color: hsl(230, 15%, 30%);
-}
-
-.titleP {
-    font-family: 'Fira Code', monospace;
-    font-size: 3rem;
-    font-weight: 700;
-}
-
-.titleS {
-    font-family: 'Fira Code', monospace;
-    font-size: 1.5rem;
-    font-weight: 700;
-}
-
-.titleT {
-    font-family: 'Fira Code', monospace;
-    font-size: 1rem;
-    font-weight: 400;
-}
-
-.logs {
-    margin: 0;
-    padding: 4rem 0;
-    position: relative;
-    list-style: none;
-}
-
-.logs li {
-    padding: 2rem 0;
-}
-
-.logs li article {
-    width: 100%;
-    display: flex;
-    align-items: stretch;
-    justify-content: flex-start;
-}
-
-.logs li:last-child article .date::before {
-    height: 100%;
-}
-
-.date {
-    word-break: keep-all;
-    padding: 0 4rem 0 0;
-    position: relative;
-    font-size: .875rem;
-}
-
-.date::before {
-    content: "";
-    width: .125rem;
-    height: calc(100% + 4rem);
-    position: absolute;
-    transform: translateX(-50%);
-    top: 0;
-    left: 100%;
-    border-radius: .125rem;
-    background-color: hsl(0, 0%, 100%);
-}
-
-.date::after {
-    content: "";
-    width: .5rem;
-    height: .5rem;
-    position: absolute;
-    transform: translateX(-50%);
-    top: 0;
-    left: 100%;
-    border: .375rem solid hsl(245, 14%, 15%);
-    border-radius: 50%;
-    background-color: hsl(0, 0%, 100%);
-}
-
-.date span {
-    position: sticky;
-    top: 0;
-}
-
-.content {
-    padding: 0 0 0 4rem;
-    display: flex;
-    flex-flow: wrap column;
-    gap: .875rem;
-}
-
-.project {
-    width: fit-content;
-    padding: .5rem;
-    line-height: 1;
-    text-transform: uppercase;
-    font-size: .625rem;
-    font-weight: 700;
-    border-radius: .25rem;
-    background: rgba(149, 128, 255, 0.16);
-}
-
-.logContent {
-    width: 100%;
-    font-size: .75rem;
-    line-height: 2;
-}
-
-.logContent img {
-    width: 100%;
-    margin: 1rem 0;
-    border-radius: .25rem;
-}
-
-.logContent a {
-    text-decoration: underline;
-    color: hsl(31, 100%, 71%);
-}
-
-.author {
-    width: 100%;
-    margin: 2rem 0 0 0;
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-}
-
-.avatar {
-    width: 1.625rem;
-    height: 1.625rem;
-    overflow: hidden;
-    border-radius: 50%;
-}
-
-.avatar img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-.authorName {
-    font-size: .75rem;
-    font-weight: 700;
-}
-
-@media (max-width: 80rem) {
-    .container {
-        width: 100%;
-    }
-
-    .logs {
-        padding: 0 10% 4rem 10%;
-    }
-
-    .logs li article {
-        position: relative;
-        flex-flow: wrap column;
-    }
-
-    .logs li article::before {
-        content: "";
-        width: .125rem;
-        height: 100%;
-        position: absolute;
-        transform: translateX(-50%);
-        top: 0;
-        left: -1.5rem;
-        border-radius: .125rem;
-        background-color: hsl(0, 0%, 100%);
-    }
-
-    .logs li article::after {
-        content: "";
-        width: .5rem;
-        height: .5rem;
-        position: absolute;
-        transform: translateX(-50%);
-        top: 0;
-        left: -1.5rem;
-        border: .375rem solid hsl(245, 14%, 15%);
-        border-radius: 50%;
-        background-color: hsl(0, 0%, 100%);
-    }
-
-    .content {
-        padding: 0;
-    }
-
-    .date {
-        padding: 0 0 1rem 0;
-    }
-
-    .date::before,
-    .date::after {
-        display: none;
-    }
-
-    .author {
-        margin: 1rem 0 0 0;
-    }
-}
